Batch Pipeline/Data_Loading/requirements.txt
psycopg2-binary
pymongo
pandas
python-dotenv
faker